% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\title{compare}
\usage{
compare(Estimate, True)
}
\arguments{
\item{Estimate}{estimated graph}

\item{True}{the true graph}
}
\value{
results in data frame
}
\description{
compare
}
