% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGMprojpred.R
\name{GGMprojpred}
\alias{GGMprojpred}
\title{GGMprojpred}
\usage{
GGMprojpred(X, n_cl, type, iter)
}
\arguments{
\item{X}{n by p data matrix}

\item{n_cl}{number of clusters for parallel}

\item{type}{regularized (using the horeshoe prior distribution) or non-regularized (Bayesian bootstrap with OLS)}

\item{iter}{number of saved posterior samples}
}
\value{

}
\description{
Estimate Gaussian graphical models with projection predictive selection
}
\examples{
GGMprojpred()
}
