% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_graph.R
\name{plot_graph}
\alias{plot_graph}
\title{plot_graph}
\usage{
plot_graph(pcor_mat, pos_col = "green", neg_col = "red", scl = 10, ...)
}
\arguments{
\item{pcor_mat}{partial correlation matrix from GGMprojpred}

\item{pos_col}{color for positive values}

\item{neg_col}{color for negative values}

\item{scl}{controls scale of edge sizes}

\item{...}{allows for passing agruments to plot.igraph}
}
\description{
Simply function for visualizing the estimated network structure
}
